Replace Device model with MyThing model
The commit represents a significant refactoring in favor of an abstracted 'MyThing' model replacing the previous 'Device' model. All instances and usages of 'Device' have been removed and replaced with 'MyThing', going as far as to change endpoint routes and update tests to match the changes. Events and handlers have also been updated to match MyThing's functionality.